Split sync repair into its own module
