+++ packages/backend/src/user/services/user-metadata.service.test.ts
@@ -8,7 +8,7 @@ import {
   setupTestDb,
 } from "@backend/__tests__/helpers/mock.db.setup";
 import { initSupertokens } from "@backend/common/middleware/supertokens.middleware";
-import syncImportRunner from "@backend/sync/services/import/sync.import-runner";
+import syncRepairRunner from "@backend/sync/services/repair/sync.repair-runner";
 import { isUsingGcalWebhookHttps } from "@backend/sync/util/sync.util";
 
 // eslint-disable-next-line @typescript-eslint/no-unsafe-return -- mock factory spreads requireActual
@@ -121,7 +121,7 @@ describe("UserMetadataService", () => {
       const user = await UserDriver.createUser();
       const userId = user._id.toString();
       const restartSpy = jest
-        .spyOn(syncImportRunner, "restartGoogleCalendarSync")
+        .spyOn(syncRepairRunner, "restartGoogleCalendarSync")
         .mockResolvedValue();
 
       const metadata = await driver.fetchUserMetadata(userId);
@@ -150,7 +150,7 @@ describe("UserMetadataService", () => {
       const user = await UserDriver.createUser();
       const userId = user._id.toString();
       const restartSpy = jest
-        .spyOn(syncImportRunner, "restartGoogleCalendarSync")
+        .spyOn(syncRepairRunner, "restartGoogleCalendarSync")
         .mockResolvedValue();
 
       await driver.updateUserMetadata({
